# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeepRef API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects/{project_id}/articles": {
      "get": {
        "tags": [
          "articles"
        ],
        "operationId": "listProjectArticles",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project articles ranked by citation metrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/articles/{doi_key}": {
      "get": {
        "tags": [
          "articles"
        ],
        "operationId": "getProjectArticle",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "doi_key",
            "in": "path",
            "description": "Base64url-encoded canonical DOI",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project article metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid DOI key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Article not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/graph": {
      "get": {
        "tags": [
          "articles"
        ],
        "operationId": "getProjectGraph",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project citation graph",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectGraphDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/recommendations": {
      "get": {
        "tags": [
          "articles"
        ],
        "operationId": "getProjectRecommendations",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project recommendation groups",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationGroupsDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleDetailDto": {
        "type": "object",
        "required": [
          "doi",
          "total_citations",
          "references_count",
          "raw"
        ],
        "properties": {
          "abstract": {
            "type": [
              "string",
              "null"
            ]
          },
          "container_title": {
            "type": [
              "string",
              "null"
            ]
          },
          "doi": {
            "type": "string"
          },
          "issued_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "published_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ]
          },
          "raw": {
            "type": "object"
          },
          "references_count": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "total_citations": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ArticleDto": {
        "type": "object",
        "required": [
          "doi",
          "doi_key",
          "total_citations",
          "internal_citations",
          "outbound_internal_references",
          "rank_score"
        ],
        "properties": {
          "doi": {
            "type": "string"
          },
          "doi_key": {
            "type": "string"
          },
          "internal_citations": {
            "type": "integer",
            "format": "int32"
          },
          "issued_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "outbound_internal_references": {
            "type": "integer",
            "format": "int32"
          },
          "rank_score": {
            "type": "number",
            "format": "double"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "total_citations": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "GraphEdgeDto": {
        "type": "object",
        "required": [
          "source",
          "target"
        ],
        "properties": {
          "source": {
            "type": "string"
          },
          "target": {
            "type": "string"
          }
        }
      },
...
      "ProjectGraphDto": {
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "properties": {
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphEdgeDto"
            }
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleDto"
            }
          }
        }
      },
      "RecommendationGroupsDto": {
        "type": "object",
        "required": [
          "foundational",
          "core_to_project",
          "underexplored"
        ],
        "properties": {
          "core_to_project": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleDto"
            }
          },
          "foundational": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleDto"
            }
          },
          "underexplored": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleDto"
            }
          }
        }
      },
...
    }
  }
}